+++ pkg/ingester/shipper.go
@@ -19,7 +19,6 @@ import (
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
-	"github.com/prometheus/prometheus/model/labels"
 	"github.com/thanos-io/thanos/pkg/block"
 	"github.com/thanos-io/thanos/pkg/block/metadata"
 	"github.com/thanos-io/thanos/pkg/objstore"
@@ -65,7 +64,6 @@ type Shipper struct {
 	dir     string
 	metrics *metrics
 	bucket  objstore.Bucket
-	labels  func() labels.Labels
 	source  metadata.SourceType
 
 	hashFunc metadata.HashFunc
@@ -79,22 +77,17 @@ func NewShipper(
 	r prometheus.Registerer,
 	dir string,
 	bucket objstore.Bucket,
-	lbls func() labels.Labels,
 	source metadata.SourceType,
 	hashFunc metadata.HashFunc,
 ) *Shipper {
 	if logger == nil {
 		logger = log.NewNopLogger()
 	}
-	if lbls == nil {
-		lbls = func() labels.Labels { return nil }
-	}
 
 	return &Shipper{
 		logger:   logger,
 		dir:      dir,
 		bucket:   bucket,
-		labels:   lbls,
 		metrics:  newMetrics(r),
 		source:   source,
 		hashFunc: hashFunc,
@@ -103,8 +96,6 @@ func NewShipper(
 
 // Sync performs a single synchronization, which ensures all non-compacted local blocks have been uploaded
 // to the object bucket once.
-//
-// If uploaded.
 //
 // It is not concurrency-safe, however it is compactor-safe (running concurrently with compactor is ok).
 func (s *Shipper) Sync(ctx context.Context) (uploaded int, err error) {
@@ -125,7 +116,7 @@ func (s *Shipper) Sync(ctx context.Context) (uploaded int, err error) {
 		hasUploaded[id] = struct{}{}
 	}
 
-	// Reset the uploaded slice so we can rebuild it only with blocks that still exist locally.
+	// Reset the uploaded slice, so we can rebuild it only with blocks that still exist locally.
 	meta.Uploaded = nil
 
 	var uploadErrs int
@@ -195,10 +186,6 @@ func (s *Shipper) upload(ctx context.Context, meta *metadata.Meta) error {
 
 	blockDir := filepath.Join(s.dir, meta.ULID.String())
 
-	// Attach current labels. Don't write the file to disk, but upload it to the bucket.
-	if lset := s.labels(); lset != nil {
-		meta.Thanos.Labels = lset.Map()
-	}
 	meta.Thanos.Source = s.source
 	meta.Thanos.SegmentFiles = block.GetSegmentFiles(blockDir)